+++ conftest.py
@@ -1,3 +1,3 @@
-from src.main.api.fixtures.api_fixtures import *
-from src.main.api.fixtures.user_fixtures import *
-from src.main.api.fixtures.object_fixtures import *
+from src.main.fixtures.api_fixtures import *
+from src.main.fixtures.user_fixtures import *
+from src.main.fixtures.object_fixtures import *
