+++ src/main/api/steps/base_steps.py
@@ -1,6 +1,10 @@
+from abc import ABC
 from typing import Any, List
 
 
-class BaseSteps:
+class BaseSteps(ABC):
     def __init__(self, created_object: List[Any]):
         self.created_objects = created_object
+
+    def add_created_object(self, created_object: Any):
+        self.created_objects.append(created_object)
