+++ src/main/api/tests/login_user_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-from src.main.api.classes.api_manager import ApiManager
+from src.main.classes.api_manager import ApiManager
 from src.main.api.models.create_user import CreateUserRequest, CreateUserResponse
