+++ src/main/configs/config.py
@@ -9,7 +9,7 @@ class Config:
     def __new__(cls):
         if cls._instance is None:
             cls._instance = super(Config, cls).__new__(cls)
-            config_path = Path(__file__).parents[4] / 'resources' / 'config.properties'
+            config_path = Path(__file__).parents[3] / 'resources' / 'config.properties'
             if not config_path.exists():
                 raise ImportError(f'Config file not found: {config_path}')
             with open(config_path, 'r') as f:
