+++ src/main/fixtures/object_fixtures.py
@@ -3,14 +3,15 @@ from typing import List, Any
 
 import pytest
 
-from src.main.api.classes.api_manager import ApiManager
 from src.main.api.models.create_user import CreateUserResponse
+from src.main.api.models.user import User
+from src.main.classes.api_manager import ApiManager
 
 
 def cleanup_objects(objects: List[Any]) -> None:
     api_manager = ApiManager(objects)
     for obj in objects:
-        if isinstance(obj, CreateUserResponse):
+        if isinstance(obj, CreateUserResponse) or isinstance(obj, User):
             api_manager.admin_steps.delete_user(obj.id)
         else:
             logging.warning(f"Object {obj.id} was not deleted")
